+++ examples/streamz/scripts/entry.sh
@@ -12,7 +12,7 @@ usage() {
     fi
     echo "Usage: $0 [POS]... [ARG]..."
     echo
-    echo "Example-1: bash $0 -b localhost:9092 -g streamz -i input -o output -m /path/to/model.pth -l /path/to/labels.yaml -p 1s -c 0,1,2 --max_batch_size 1000 --benchmark"
+    echo "Example-1: bash $0 -b localhost:9092 -g streamz -i input -o output -m /path/to/model.pth -l /path/to/labels.yaml -p 1s -c 0,1,2 --max_batch_size 1000 --benchmark 20.0"
     echo
     echo "Run cybert model using kafka"
     echo
@@ -26,7 +26,7 @@ usage() {
     echo "  -d, --data                 Cybert data file (optional)"
     echo "  -p, --poll_interval        Poll interval (ex:60s)"
     echo "  -s, --max_batch_size       Max batch size"
-    echo "  --benchmark                Benchmark cyBERT data processing (optional)"
+    echo "  --benchmark                Benchmark cyBERT data processing. Please enter the avg log size in kb for throughput estimate. (optional)"
     echo
     echo "  -h, --help          Print this help"
     echo
@@ -55,7 +55,7 @@ while [ $# != 0 ]; do
     -d|--data) shift; data=$1 ;;
     -p|--poll_interval) shift; poll_interval=$1;;
     -s|--max_batch_size) shift; max_batch_size=$1;;
-    --benchmark) benchmark="true" ;;
+    --benchmark) shift; benchmark=$1;;
     -) usage "Unknown positional: $1" ;;
     -?*) usage "Unknown positional: $1" ;;
     esac