+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 - PR #232 Update phishing detection to use cudf subword tokenizer
 - PR #237 Cybert streamz memory optimization
 - PR #238 Deprecate CLX subword tokenizer
+- PR #239 Added cybert streamz log size parameter
 - PR #244 Add cybert dataloader
+- PR #247 Allow CuPy 8.x
 
 ## Bug Fixes
 - PR #231 Fix segmentation fault in cybert notebook